fix: critical crash and robustness audit — auth, sync, offline, attendance, FO screens
Auth & Session:
- Add dispose() override to AuthSessionNotifier to cancel Firebase subscription
- Await subscription cancellation to prevent duplicate listeners
- Add _disposed guard to prevent post-disposal state writes
- Fix unsafe forced unwraps on _cachedSession class field
- Fix auth_gate error state to include Retry and Sign-out buttons
- Add mounted guards and error feedback to biometric check
- Fix permission check error handling and reset on session change
- Fix _authHeaders() to throw when unauthenticated instead of sending empty headers
- Fix resolveCurrentSession empty catch to log errors

Network:
- Fix _getJson/_postJson/_patchJson to validate response.data is a Map before casting
- Fix _isOfflineDioError to include unknown and badCertificate types
- Fix _getJson to rethrow offline errors (was swallowing them)

Sync:
- Fix SyncService.start() to call processQueue() immediately on startup
- Add retry eviction after 20 failed attempts (was no-op)
- Add try/catch around queue removal to prevent duplicate submissions
- Validate photo upload URLs before using them
- Fix FO report upload path to use currentUser?.uid instead of !
- Add response success validation ({success: false} rejection)

Offline Queue:
- Fix getEncryptionCipher to catch secure storage/base64 errors
- Fix getQueuedRequests to skip corrupted entries instead of crashing entire queue
- Fix OfflineRequest.fromJson to handle null/malformed fields gracefully

Attendance:
- Fix QR flow _employee! and _selectedSite! forced unwraps with null checks
- Fix guard attendance GPS to use 10s timeout with proper error feedback
- Fix guard attendance silent return when position is null

FO Screens:
- Fix reports screen TextEditingController recreation on every build
- Fix attendance screen firstWithout orElse crash when selected site missing
- Fix work_orders shared ScrollController across TabBarView tabs
- Fix reports screen FirebaseAuth.currentUser!.uid crash (2 locations)
- Fix dashboard .clamp() missing .toDouble() (3 locations)

All changes pass flutter analyze.

+++ lib/features/field_officer/presentation/screens/field_officer_dashboard_screen.dart
@@ -240,7 +240,7 @@ class _DeploymentTile extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final tokens = CissThemeTokens.of(context);
-    final double progress = total <= 0 ? 0.0 : (active / total).clamp(0, 1);
+    final double progress = total <= 0 ? 0.0 : (active / total).clamp(0, 1).toDouble();
     final int percent = (progress * 100).toInt();
 
     return GlassCard(
@@ -624,7 +624,7 @@ class _ComplianceTile extends StatelessWidget {
   Widget build(BuildContext context) {
     final tokens = CissThemeTokens.of(context);
     final double progress =
-        onDuty <= 0 ? 0.0 : (checkedIn / onDuty).clamp(0, 1);
+        onDuty <= 0 ? 0.0 : (checkedIn / onDuty).clamp(0, 1).toDouble();
     final int percent = (progress * 100).toInt();
 
     return GlassCard(
